requirements.txt
colorama==0.4.6
pycryptodome==3.23.0
tqdm==4.67.1
